Refactor Users page by extracting TextFilterDropdown component for improved code organization and maintainability, and remove inline filter implementation from the Users component.
